backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
chromadb
pypdf
python-dotenv
numpy
